Revert "feat: added cli logging to native [DET-3316] (#788)"
This reverts commit 0c99a0cda9850543ee796c3e660c2c29f74c3c6a.
