+++ cli/determined_cli/agent.py
@@ -7,10 +7,10 @@ from typing import Any, Callable, List
 
 from determined_cli import render
 from determined_common import api
-from determined_common.api.authentication import authentication_required
 from determined_common.check import check_false
 
 from .declarative_argparse import Arg, Cmd, Group
+from .user import authentication_required
 
 
 def local_id(address: str) -> str:
