+++ cli/determined_cli/remote.py
@@ -6,7 +6,6 @@ from typing import Any, Dict, List
 from termcolor import colored
 
 from determined_common import api
-from determined_common.api.authentication import authentication_required
 
 from . import render
 from .command import (
@@ -21,6 +20,7 @@ from .command import (
     render_event_stream,
 )
 from .declarative_argparse import Arg, Cmd
+from .user import authentication_required
 
 
 @authentication_required
