+++ cli/determined_cli/shell.py
@@ -8,7 +8,6 @@ from typing import Any, Dict, List
 from termcolor import colored
 
 from determined_common import api
-from determined_common.api.authentication import authentication_required
 from determined_common.check import check_eq, check_len
 
 from . import render
@@ -24,6 +23,7 @@ from .command import (
     render_event_stream,
 )
 from .declarative_argparse import Arg, Cmd
+from .user import authentication_required
 
 
 def get_agent_user(host: str) -> str:
