+++ cli/determined_cli/tensorboard.py
@@ -6,12 +6,12 @@ from typing import Any, Dict, List
 from termcolor import colored
 
 from determined_common import api
-from determined_common.api.authentication import authentication_required
 from determined_common.check import check_eq
 
 from . import render
 from .command import Command, render_event_stream
 from .declarative_argparse import Arg, Cmd
+from .user import authentication_required
 
 Tensorboard = namedtuple(
     "Tensorboard",
