+++ cli/determined_cli/trial.py
@@ -1,15 +1,18 @@
 import distutils.util
 import json
+import time
 from argparse import Namespace
-from typing import Any, List
+from typing import Any, List, Optional
+
+from termcolor import colored
 
 from determined_cli import render
-from determined_common import api
-from determined_common.api.authentication import authentication_required
+from determined_common import api, constants
 from determined_common.experimental import Determined
 
 from .checkpoint import format_checkpoint, format_validation, render_checkpoint
 from .declarative_argparse import Arg, Cmd, Group
+from .user import authentication_required
 
 
 @authentication_required
@@ -75,6 +78,52 @@ def describe_trial(args: Namespace) -> None:
     print()
     print("Steps:")
     render.tabulate_or_csv(headers, values, args.csv)
+
+
+@authentication_required
+def logs(args: Namespace) -> None:
+    last_offset, last_state = 0, None
+
+    def print_logs(offset: Optional[int], limit: Optional[int] = 5000) -> Any:
+        nonlocal last_offset, last_state
+        path = "trials/{}/logsv2?".format(args.trial_id)
+        if offset is not None:
+            path += "&offset={}".format(offset)
+        if limit is not None:
+            path += "&limit={}".format(limit)
+        logs = api.get(args.master, path).json()
+        for log in logs:
+            print(log["message"], end="")
+            last_state = log["state"]
+        return logs[-1]["id"] if logs else last_offset
+
+    try:
+        if args.tail is not None:
+            last_offset = print_logs(None, args.tail)
+        else:
+            while True:
+                new_offset = print_logs(last_offset)
+                if last_offset == new_offset:
+                    break
+                last_offset = new_offset
+
+        if not args.follow:
+            return
+        while True:
+            last_offset = print_logs(last_offset)
+            if last_state in constants.TERMINAL_STATES:
+                break
+            time.sleep(0.2)
+    except KeyboardInterrupt:
+        pass
+    finally:
+        print(
+            colored(
+                "Trial is in the {} state. To reopen log stream, run: "
+                "det trial logs -f {}".format(last_state, args.trial_id),
+                "green",
+            )
+        )
 
 
 def download(args: Namespace) -> None:
@@ -179,7 +228,7 @@ args_description = [
             ),
             Cmd(
                 "logs",
-                api.experiment.logs,
+                logs,
                 "fetch trial logs",
                 [
                     Arg("trial_id", type=int, help="trial ID"),
